requirements.txt
certifi==2018.4.16
chardet==3.0.4
clearbit==0.1.7
Django==2.0.6
django-crispy-forms==1.7.2
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.6
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
requests==2.18.4
sorl-thumbnail==12.4.1
urllib3==1.22
